requirements.txt
django==4.1
django-filter==23.2
psycopg2>=2.8
djangorestframework==3.14.0
drf-yasg==1.21.5
celery>5.2
redis==4.3.4
requests==2.28